+++ actix-identity/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Use `std::future::{ready, Ready}` instead of `actix-utils` types.
 - Minimum supported Rust version (MSRV) is now 1.88.
 
 ## 0.9.0