+++ actix-web-httpauth/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Use `std::future::{ready, Ready}` instead of `actix-utils` types.
+
 ## 0.8.3
 
 - Minimum supported Rust version (MSRV) is now 1.88.